refactor(useInput): useInput을 일반적인 함수로 만들고 사용하는 커스텀 훅에서 적절하게 이벤트 핸들러 처리하기
